#1 Added error messages and exception handling in user controller

+++ backend/tracker/src/main/java/com/TrackYourGains/tracker/controller/UserController.java
@@ -1,8 +1,10 @@
 package com.TrackYourGains.tracker.controller;
 
 import com.TrackYourGains.tracker.dto.UserDto;
+import com.TrackYourGains.tracker.repository.UserRepository;
 import com.TrackYourGains.tracker.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,6 +15,9 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private UserRepository userRepository;
+
     /**
      * Endpoint that returns a user detail
      * @param id id of the user in database
@@ -22,6 +27,8 @@ public class UserController {
     public ResponseEntity<UserDto> getUserByID(@PathVariable Long id)
     {
         UserDto user = userService.getUserById(id);
+        if(user == null)
+            return new ResponseEntity("User not found", HttpStatus.NOT_FOUND);
         return ResponseEntity.ok(user);
     }
 
@@ -33,6 +40,8 @@ public class UserController {
     @PostMapping
     public ResponseEntity<UserDto> createUser(@RequestBody UserDto userDto)
     {
+        if(!userRepository.existsByFirstNameAndLastName(userDto.getFirstName(), userDto.getLastName()))
+            return new ResponseEntity("User does not exist", HttpStatus.BAD_REQUEST);
         UserDto user = userService.addUser(userDto);
         return ResponseEntity.ok(user);
     }
@@ -46,6 +55,8 @@ public class UserController {
     @PutMapping("/{id}")
     public ResponseEntity<UserDto> updateUserDetails(@PathVariable Long id, @RequestBody UserDto userDto)
     {
+        if(!userRepository.existsById(id))
+            return new ResponseEntity("User does not exist", HttpStatus.BAD_REQUEST);
         UserDto user = userService.updateUser(id, userDto);
         return ResponseEntity.ok(user);
     }
@@ -58,6 +69,8 @@ public class UserController {
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteUser(@PathVariable Long id)
     {
+        if(!userRepository.existsById(id))
+            return new ResponseEntity("User does not exist",  HttpStatus.BAD_REQUEST);
         userService.deleteUser(id);
         return ResponseEntity.ok("Deleted user successfully!");
     }
